+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 1.4.13
 - Fixed Portable Grid model (raoulvdberge, CyanideX)
+- Fixed ore dictionary causing problems with Solderer (raoulvdberge)
+- Fixed ore dictionary items not showing up in JEI for the Solderer (raoulvdberge)
 
 ### 1.4.12
 - Updated Forge to 2315 (raoulvdberge)